RD-4188 Move components_constants & service_names
I don't think there's a good reason to keep those under components/
(indeed, why would they be called "component_constants" if they're
in components/, and not just "constants"?)

But there is a good reason to keep them outside! It avoids a circular
import between utils/common.py and components/__init__.py
