+++ cfy_manager/components/__init__.py
@@ -32,7 +32,7 @@ from .stage import Stage  # NOQA
 from .usage_collector import UsageCollector  # NOQA
 from .execution_scheduler import ExecutionScheduler  # NOQA
 from .rsyslog import Rsyslog  # NOQA
-from .service_names import (  # NOQA
+from ..service_names import (  # NOQA
     DATABASE_SERVICE,
     ENTROPY_SERVICE,
     MANAGER_SERVICE,
