+++ cfy_manager/components/stage/stage.py
@@ -2,7 +2,7 @@ import os
 import json
 from os.path import join, isfile
 
-from ..components_constants import (
+from ...components_constants import (
     CLUSTER_JOIN,
     PRIVATE_IP,
     SSL_CLIENT_VERIFICATION,
@@ -10,7 +10,7 @@ from ..components_constants import (
     SSL_INPUTS,
 )
 from ..base_component import BaseComponent
-from ..service_names import (
+from ...service_names import (
     MANAGER,
     POSTGRESQL_CLIENT,
     STAGE,
