+++ cfy_manager/components_constants.py
