+++ cfy_manager/utils/common.py
@@ -18,8 +18,8 @@ from ..config import config
 from ..logger import get_logger
 from ..exceptions import ProcessExecutionError
 
-from cfy_manager.components.components_constants import SERVICES_TO_INSTALL
-from cfy_manager.components.service_names import (
+from cfy_manager.components_constants import SERVICES_TO_INSTALL
+from cfy_manager.service_names import (
     POSTGRESQL_SERVER,
     QUEUE_SERVICE,
     MANAGER_SERVICE,
